* nfacctd: introduced support for 16 bits SAMPLER_IDs in NetFlow v9/IPFIX;   this appears to be the standard length with IOS-XR. * nfacctd: introduced support for (FLOW)_SAMPLING_INTERVAL fields as part   of the NetFlow v9/IPFIX data record. This case is not prevented by the   RFC although such information is typically exported as part of an option.   It appears some probes, ie. FlowMon by Invea-Tech, are getting down this   way.

+++ src/xflow_status.c
@@ -188,7 +188,7 @@ search_smp_if_status_table(struct xflow_status_entry_sampling *sentry, u_int32_t
 }
 
 struct xflow_status_entry_sampling *
-search_smp_id_status_table(struct xflow_status_entry_sampling *sentry, u_int8_t sampler_id, u_int8_t return_unequal)
+search_smp_id_status_table(struct xflow_status_entry_sampling *sentry, u_int16_t sampler_id, u_int8_t return_unequal)
 {
   /* Match a samplerID or, if samplerID within a data record is zero and no match was
      possible, then return the last samplerID defined -- last part is C7600 workaround */
